backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
passlib[bcrypt]
python-jose
langdetect
pillow
opencv-python
transformers
torch
sentence-transformers
pymupdf
easyocr